+++ CHANGELOG.md
@@ -177,7 +177,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 - Drop the hook system for modifying the CSV export (#5134)
-- Drop the legacy FE registration lists (#4277, #5133)
+- Drop the legacy FE registration lists (#4277, #5133, #5137)
 - Drop the email hook system (#5085)
 - Drop the CSV attachment from the scheduler task emails (#5028)
 - Drop `LegacyEvent::isUserRegisteredMessage()` (#4985)